Reorganization of publication related clasees. [noissue]
